Ex01 done

+++ ex01/PhoneBook.class.hpp
@@ -2,7 +2,7 @@
 # define PHONE_BOOK_HPP
 
 #include <iostream>
-#include "Contact.class.hpp"
+#include <Contact.class.hpp>
 
 class PhoneBook {
 
@@ -11,16 +11,16 @@ public:
 	PhoneBook(const PhoneBook &copy);
 	~PhoneBook(void);
 	PhoneBook& operator= (const PhoneBook &copy);
-	int		get_n(void) const;
-	void	set_n(const int n);
+	size_t	get_n(void) const;
+	void	set_n(const size_t n);
 	void	add(void);
 	void	search(void);
 	int		full(void);
 
 private:
-	const static int	max = 8;
+	const static size_t	max = 8;
 	Contact				contacts[max];
-	int					n;
+	size_t				n;
 };
 
 #endif
